Update STALL fix to not change timing
The check for a STALL condition on an OUT endpoint that has a buffer
ready now occurs after receving the data packet. This check replaces
several NOPs that were present to ensure a minimum interpacket delay.
This check should therefore not change the timing.

The data packet will be received into the buffer provided by the
application, but if the endpoint is halted, then a STALL will be issued,
the application will not be notified, and the ready condition will not
be cleared.