STALL halted IN EP even when marked ready
This allows IN endpoints to be STALLed when they are halted, even if
the application has prepare a buffer for it. This is similar to the fix
for OUT endpoints.

This introduces more cycles between receving an IN token packet, and
replying with a DATA packet, but this should still be well under the
maximum time allowed by the USB spec.

This case is seems less important than it is for OUT endpoints, but I do
think we should handle it for completeness.